trying to simplify code
