--- src/human/comparator/HumanComparatorByBirthDate.java
@@ -1,11 +1,0 @@
-package human.comparator;
-
-import human.Human;
-import member.comparator.MemberComparator;
-
-public class HumanComparatorByBirthDate implements MemberComparator<Human> {
-    @Override
-    public int compare(Human h1, Human h2) {
-        return h1.getBirthDate().compareTo(h2.getBirthDate());
-    }
-}
